pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "deskflow-agent"
version = "1.0.0"
description = "LangGraph IT support agent for DeskFlow Inc."
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "chromadb>=0.5.0",
    "openai>=1.40.0",
    "httpx>=0.27.0",
    "motor>=3.5.0",
    "pymongo>=4.8.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["deskflow_agent"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
